Refactor a bunch of class names. Added documentation
