--- src/main/java/com/sirnommington/interception/Interceptor.java
@@ -1,5 +1,0 @@
-package com.sirnommington.interception;
-
-public interface Interceptor {
-    Object execute(InterceptorOperationContext operation);
-}
