--- src/main/java/com/sirnommington/interception/InterceptorOperationContext.java
@@ -1,7 +1,0 @@
-package com.sirnommington.interception;
-
-public interface InterceptorOperationContext {
-    String getOperationName();
-    Object getInput();
-    Object execute();
-}
