+++ src/main/java/com/sirnommington/interception/core/Operation.java
@@ -1,5 +1,8 @@
-package com.sirnommington.interception;
+package com.sirnommington.interception.core;
 
+import com.sirnommington.interception.ExecutableOperation;
+import com.sirnommington.interception.interceptor.Interceptor;
+import com.sirnommington.interception.interceptor.ContinuableOperation;
 import lombok.Setter;
 import lombok.experimental.Accessors;
 
@@ -9,7 +12,7 @@ import java.util.function.Function;
 import java.util.function.Supplier;
 
 @Accessors(fluent = true)
-public class Operation implements InterceptorOperationContext, ExecutableOperation {
+public class Operation implements ContinuableOperation, ExecutableOperation {
     private final Iterator<Interceptor> interceptors;
 
     @Setter
@@ -18,7 +21,7 @@ public class Operation implements InterceptorOperationContext, ExecutableOperati
     private Object input;
     private Function<Object, Object> func;
 
-    Operation(Iterator<Interceptor> interceptors) {
+    public Operation(Iterator<Interceptor> interceptors) {
         this.interceptors = interceptors;
     }
 
@@ -80,6 +83,6 @@ public class Operation implements InterceptorOperationContext, ExecutableOperati
         this.input = input;
         this.func = (objInput) -> func.apply((T) objInput);
 
-        return (R)execute();
+        return (R) execute();
     }
 }
